PINN/requirements.txt
sympy>=1.14
torch>=2.2
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
pix2tex[api]>=0.1.4
antlr4-python3-runtime>=4.11